Updated-loadingFXMLfiles-paths & new User obj for every login
I updated the FXML file paths for the scene controller as richie updated the file paths for the back button to work etc. Added getFromDatabase func to use sqlite to get the user's info and put the info in a User object thats created everytime a user logins. So we can pass that User object into a controller for scene builder.